requirements.txt
streamlit
pandas
requests
python-dotenv
numpy
# clingo (install via conda if pip install fails)